Create coredump group of commands to handle corefile management
