+++ config/main.py
@@ -27,6 +27,7 @@ from .utils import log
 from . import aaa
 from . import chassis_modules
 from . import console
+from . import coredump
 from . import feature
 from . import kube
 from . import mlnx
@@ -877,6 +878,7 @@ config.add_command(aaa.aaa)
 config.add_command(aaa.tacacs)
 config.add_command(chassis_modules.chassis_modules)
 config.add_command(console.console)
+config.add_command(coredump.coredump)
 config.add_command(feature.feature)
 config.add_command(kube.kubernetes)
 config.add_command(muxcable.muxcable)
@@ -1900,33 +1902,6 @@ def bgp():
 def shutdown():
     """Shut down BGP session(s)"""
     pass
-
-@config.group(cls=clicommon.AbbreviationGroup)
-def core():
-    """ Configure coredump """
-    if os.geteuid() != 0:
-        exit("Root privileges are required for this operation")
-    pass
-
-@core.command()
-@click.argument('disable', required=False)
-def disable(disable):
-    """Administratively Disable coredump generation"""
-    config_db = ConfigDBConnector()
-    config_db.connect()
-    table = "COREDUMP"
-    key = "config"
-    config_db.set_entry(table, key, {"enabled": "false"})
-
-@core.command()
-@click.argument('enable', required=False)
-def enable(enable):
-    """Administratively Enable coredump generation"""
-    config_db = ConfigDBConnector()
-    config_db.connect()
-    table = "COREDUMP"
-    key = "config"
-    config_db.set_entry(table, key, {"enabled": "true"})
 
 @config.group(cls=clicommon.AbbreviationGroup)
 def kdump():
